+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where Studio .io file custom parts would not import in some cases due to incorrect path handling.
+
 ## 0.5.0 - 2026-03-17
 ### Changed
 * Changed supported Blender version to 5.1 or later.